Rewrite register system
